Added Wireframes
Includes the full wireframe called Home Page.png, but since this is so big it might be hard to see, so i also captured the wireframes for the flow of the 3 main functionalities. vieweditwireframe.png is for the View/Edit Team functionality, CreateTeamWireFrame.png is for the creating a team functionality. ViewOthersTeamsWireframe.png is for the functionality of viewing other users pokemon teams.